# docs/api/openapi-b2b-workflow.yaml
openapi: 3.0.3
info:
  title: EPACKAGE Lab B2B Workflow API
  description: |
    10단계 B2B 주문 프로세스를 위한 RESTful API

    워크플로우:
    1. Registration (회원가입)
    2. Quotation (견적)
    3. Order (주문)
    4. Data Entry (데이터입고)
    5. Work Order (작업표준서)
    6. Contract (계약서)
    7. Signing (전자서명)
    8. Production (생산)
    9. Stock In (입고)
    10. Shipment (출하)
  version: 1.0.0
  contact:
    name: EPACKAGE Lab
    email: api@epackage-lab.com
servers:
  - url: https://api.epackage-lab.com/api/v1
    description: Production server
  - url: https://staging-api.epackage-lab.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /quotations:
    post:
      tags: [Quotations]
      summary: 견적 생성
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuotationRequest'
      responses:
        '201':
          description: 견적 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateQuotationResponse'
    get:
      tags: [Quotations]
      summary: 견적 목록 조회
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/CompanyIdParam'
        - $ref: '#/components/parameters/FromDateParam'
        - $ref: '#/components/parameters/ToDateParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListQuotationsResponse'
  /quotations/{id}:
    get:
      tags: [Quotations]
      summary: 견적 상세 조회
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetQuotationResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 토큰 인증
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    StatusParam:
      name: status
      in: query
      schema:
        type: string
        enum: [DRAFT, SENT, APPROVED, REJECTED, EXPIRED, CONVERTED]
...
    UserIdParam:
      name: user_id
      in: query
      schema:
        type: string
    CompanyIdParam:
      name: company_id
      in: query
      schema:
        type: string
    FromDateParam:
      name: from_date
      in: query
      schema:
        type: string
        format: date
    ToDateParam:
      name: to_date
      in: query
      schema:
        type: string
        format: date
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        enum: [created_at, total_amount, valid_until]
    OrderParam:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
...
  schemas:
...
    CreateQuotationRequest:
      type: object
      required:
        - customer_name
        - customer_email
        - items
      properties:
        customer_name:
          type: string
        customer_email:
          type: string
        customer_phone:
          type: string
        items:
          type: array
          items:
            type: object
            required:
              - product_name
              - quantity
              - unit_price
            properties:
              product_name:
                type: string
              product_code:
                type: string
              category:
                type: string
              quantity:
                type: integer
                minimum: 1
              unit_price:
                type: number
                minimum: 0
              specifications:
                type: object
              notes:
                type: string
        valid_days:
          type: integer
          minimum: 7
          maximum: 90
        notes:
          type: string
    CreateQuotationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            quotation_number:
              type: string
            status:
              type: string
            subtotal_amount:
              type: number
            tax_amount:
              type: number
            total_amount:
              type: number
            valid_until:
              type: string
            created_at:
              type: string
...